improve: migrate RingBuffer to TypeScript and add accessors
